organize into seperate components
