+++ src/App.js
@@ -1,100 +1,12 @@
-import React, { Component } from "react";
-import "./App.css";
+import React from 'react'
+import Test from './components/InfiniteScroll.2'
 
-class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      error: null,
-      loading: false,
-      data: [],
-      perPage: 20,
-      page: 1
-    };
-  }
-
-  componentDidMount() {
-    this.loadingData();
-    window.addEventListener("scroll", e => {
-      if (!this.state.loading) {
-        this.infiniteScroll();
-      }
-    });
-  }
-
-  getData() {
-    const { page, perPage } = this.state;
-    fetch(
-      `https://jsonplaceholder.typicode.com/comments?_page=${page}&_limit=${perPage}`
-    )
-      .then(response => response.json())
-      .then(response => {
-        this.setState(prevState => ({
-          page: prevState.page + 1,
-          loading: false,
-          data: [...this.state.data, ...response]
-        }));
-      })
-      .catch(error => {
-        this.setState({
-          error
-        });
-        console.log("error: ", error);
-      });
-  }
-
-  infiniteScroll() {
-    // checks to see if div is scrolled into view
-    let lastElement = document.getElementsByClassName("lastElement");
-    let bounding = lastElement[0].getBoundingClientRect();
-    let elementTop = bounding.top;
-    let elementBottom = bounding.bottom;
-    let isVisible = elementTop >= 0 && elementBottom <= window.innerHeight;
-    if (isVisible) {
-      this.loadingData();
-    }
-  }
-
-  loadingData() {
-    this.setState({ loading: true });
-    this.getData();
-  }
-
-  render() {
-    const { error, loading, data } = this.state;
-    if (error) {
-      return <div>Error: {error.message}</div>;
-    } else if (loading) {
-      return <div className="loader"></div>;
-    } else {
-      return (
-        <div className="App">
-          <h1>Infinite Scroll Challenge</h1>
-          <div className="cards">
-            {data.map(item => (
-              <div key={item.id} className="card">
-                <div className="info">
-                  <div className="email">{item.email}</div>
-                  <div className="name">{item.name}</div>
-                  <div className="comment">{item.body}</div>
-                </div>
-              </div>  
-            ))}
-          </div>
-          <div className="lastElement">
-            <button
-              onClick={e => {
-                this.loadingData();
-              }}
-              className="button"
-            >
-              load more
-            </button>
-          </div>
-        </div>
-      );
-    }
-  }
+const App = () => {
+  return (
+    <div>
+      <Test />
+    </div>
+  )
 }
 
-export default App;
+export default App
